+++ samples/Avalonia/Demo.SaveFileDialog/MainWindow.axaml.cs
@@ -1,3 +1,5 @@
+using System;
+using System.Threading.Tasks;
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
@@ -18,4 +20,22 @@ public partial class MainWindow : Window
     {
         AvaloniaXamlLoader.Load(this);
     }
+    //
+    // protected override async void OnOpened(EventArgs e)
+    // {
+    //     var result = await ShowSaveFileAsync();
+    //     base.OnOpened(e);
+    // }
+    //
+    // private async Task<string?> ShowStrAsync() =>
+    //     (string?)await ShowObjAsync();
+    //
+    // private async Task<object?> ShowObjAsync() =>
+    //     await ShowSaveFileAsync();
+    //
+    // private async Task<string?> ShowSaveFileAsync()
+    // {
+    //     var dialog = new Avalonia.Controls.SaveFileDialog();
+    //     return await dialog.ShowAsync(this).ConfigureAwait(true);
+    // }
 }