+++ src/MvvmDialogs.Avalonia/DialogFactory.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using HanumanInstitute.MvvmDialogs.Avalonia.Api;
 using HanumanInstitute.MvvmDialogs.FrameworkDialogs;
@@ -39,10 +40,10 @@ public class DialogFactory : DialogFactoryBase
     public override async Task<object?> ShowDialogAsync<TSettings>(WindowWrapper owner, TSettings settings, AppDialogSettings appSettings) =>
         settings switch
         {
-            OpenFolderDialogSettings s => await ShowOpenFolderDialogAsync(owner, s, appSettings),
-            OpenFileDialogSettings s => await ShowOpenFileDialogAsync(owner, s, appSettings),
-            SaveFileDialogSettings s => await ShowSaveFileDialogAsync(owner, s, appSettings),
-            _ => base.ShowDialogAsync(owner, settings, appSettings)
+            OpenFolderDialogSettings s => await ShowOpenFolderDialogAsync(owner, s, appSettings).ConfigureAwait(true),
+            OpenFileDialogSettings s => await ShowOpenFileDialogAsync(owner, s, appSettings).ConfigureAwait(true),
+            SaveFileDialogSettings s => await ShowSaveFileDialogAsync(owner, s, appSettings).ConfigureAwait(true),
+            _ => await base.ShowDialogAsync(owner, settings, appSettings).ConfigureAwait(true)
         };
 
     private async Task<string?> ShowOpenFolderDialogAsync(WindowWrapper owner, OpenFolderDialogSettings settings, AppDialogSettings appSettings)
@@ -54,7 +55,7 @@ public class DialogFactory : DialogFactoryBase
             // d.ShowNewFolderButton = Settings.ShowNewFolderButton;
         };
 
-        return await _api.ShowOpenFolderDialog(owner.Ref, apiSettings).ConfigureAwait(false);
+        return await _api.ShowOpenFolderDialogAsync(owner.Ref, apiSettings).ConfigureAwait(true);
     }
 
     private async Task<string[]> ShowOpenFileDialogAsync(WindowWrapper owner, OpenFileDialogSettings settings, AppDialogSettings appSettings)
@@ -67,7 +68,27 @@ public class DialogFactory : DialogFactoryBase
         };
         AddSharedSettings(apiSettings, settings);
 
-        return await _api.ShowOpenFileDialog(owner.Ref, apiSettings).ConfigureAwait(false) ?? Array.Empty<string>();
+        string[] result;
+        bool resultValid;
+        do
+        {
+            resultValid = true;
+            result = await _api.ShowOpenFileDialogAsync(owner.Ref, apiSettings).ConfigureAwait(true) ?? Array.Empty<string>();
+            for (var i = 0; i < result.Length; i++)
+            {
+                // Work around the limitation that async method cannot return ref. We may add DefaultExtension to the result.
+                var refItem = new RefClass<string>(result[i]);
+                resultValid = await CheckResultAsync(owner, refItem, settings, appSettings).ConfigureAwait(true);
+                result[i] = refItem.Value;
+                if (!resultValid)
+                {
+                    apiSettings.InitialFileName = result[i];
+                    break;
+                }
+            }
+        } while (!resultValid);
+
+        return result;
     }
 
     private async Task<string?> ShowSaveFileDialogAsync(WindowWrapper owner, SaveFileDialogSettings settings, AppDialogSettings appSettings)
@@ -75,21 +96,35 @@ public class DialogFactory : DialogFactoryBase
         var apiSettings = new SaveFileApiSettings()
         {
             DefaultExtension = settings.DefaultExtension
-            // d.CreatePrompt = Settings.CreatePrompt;
-            // d.OverwritePrompt = Settings.OverwritePrompt;
         };
         AddSharedSettings(apiSettings, settings);
 
-        var result = await _api.ShowSaveFileDialog(owner.Ref, apiSettings).ConfigureAwait(false);
+        string? result;
+        bool resultValid;
+        do
+        {
+            resultValid = true;
+            result = await _api.ShowSaveFileDialogAsync(owner.Ref, apiSettings).ConfigureAwait(true);
+            if (result != null)
+            {
+                // Work around the limitation that async method cannot return ref. We may add DefaultExtension to the result.
+                var refItem = new RefClass<string>(result);
+                resultValid = await CheckResultAsync(owner, refItem, settings, appSettings).ConfigureAwait(true);
+                result = refItem.Value;
+                if (!resultValid)
+                {
+                    apiSettings.InitialFileName = result;
+                    apiSettings.Directory = Path.GetDirectoryName(result);
+                }
+            }
+        } while (!resultValid);
+
         return result;
     }
 
     private void AddSharedSettings(FileApiSettings d, FileDialogSettings s)
     {
-        // s.DefaultExtension
         // d.DereferenceLinks = s.DereferenceLinks;
-        // d.CheckFileExists = s.CheckFileExists;
-        // d.CheckPathExists = s.CheckPathExists;
         d.Directory = s.InitialDirectory;
         d.InitialFileName = s.InitialFile;
         d.Filters = SyncFilters(s.Filters);
@@ -101,6 +136,96 @@ public class DialogFactory : DialogFactoryBase
             x => new FileDialogFilter()
             {
                 Name = x.NameToString(x.ExtensionsToString()),
-                Extensions = x.Extensions
+                Extensions = x.Extensions.Select(y => y.TrimStart('.')).ToList()
             }).ToList();
+
+    private async Task<bool> CheckResultAsync(WindowWrapper owner, RefClass<string> value, FileDialogSettings settings, AppDialogSettings appSettings)
+    {
+        var fileInfo = _pathInfo.GetFileInfo(value.Value);
+
+        // DefaultExtension.
+        if (!string.IsNullOrEmpty(settings.DefaultExtension) && !fileInfo.Exists && !value.Value.Contains('.'))
+        {
+            value.Value += "." + settings.DefaultExtension.TrimStart('.');
+            fileInfo = _pathInfo.GetFileInfo(value.Value);
+        }
+
+        // CheckFileExists.
+        if (settings.CheckFileExists && !fileInfo.Exists)
+        {
+            var msgSettings = new MessageBoxSettings()
+            {
+                Title = "File Not Found",
+                Text = "Selected file does not exist:" + Environment.NewLine + value.Value,
+                Icon = MessageBoxImage.Exclamation
+            };
+            await ChainTop.ShowDialogAsync(owner, msgSettings, appSettings).ConfigureAwait(true);
+            return false;
+        }
+
+        // CheckPathExists.
+        var path = Path.GetDirectoryName(value.Value)!;
+        var pathInfo = _pathInfo.GetDirectoryInfo(path);
+        if (settings.CheckPathExists && !pathInfo.Exists)
+        {
+            var msgSettings = new MessageBoxSettings()
+            {
+                Title = "Folder Not Found",
+                Text = "Selected folder does not exist." + Environment.NewLine + path,
+                Icon = MessageBoxImage.Exclamation
+            };
+            await ChainTop.ShowDialogAsync(owner, msgSettings, appSettings).ConfigureAwait(true);
+            return false;
+        }
+
+        if (settings is SaveFileDialogSettings saveSettings)
+        {
+            // CreatePrompt.
+            if (saveSettings.CreatePrompt && !fileInfo.Exists)
+            {
+                var msgSettings = new MessageBoxSettings()
+                {
+                    Title = "Create Confirmation",
+                    Text = "File doesn't exist, do you want to create it?" + Environment.NewLine + value.Value,
+                    Button = MessageBoxButton.YesNo,
+                    DefaultValue = true
+                };
+                var result = (bool?)await ChainTop.ShowDialogAsync(owner, msgSettings, appSettings).ConfigureAwait(true);
+                if (result != true)
+                {
+                    return false;
+                }
+            }
+
+            // OverwritePrompt.
+            if (saveSettings.OverwritePrompt && fileInfo.Exists)
+            {
+                var msgSettings = new MessageBoxSettings()
+                {
+                    Title = "Overwrite Confirmation",
+                    Text = "File already exists, do you want to overwrite it?" + Environment.NewLine + value.Value,
+                    Button = MessageBoxButton.YesNo,
+                    DefaultValue = true
+                };
+                var result = (bool?)await ChainTop.ShowDialogAsync(owner, msgSettings, appSettings).ConfigureAwait(true);
+                if (result != true)
+                {
+                    return false;
+                }
+            }
+        }
+
+        return true;
+    }
+
+    /// <summary>
+    /// Works around the limitation that async method cannot return ref parameter. Classes are by reference. 
+    /// </summary>
+    /// <typeparam name="T">The data type to pass by reference.</typeparam>
+    private class RefClass<T>
+    {
+        public RefClass(T value) => Value = value;
+
+        public T Value { get; set; }
+    }
 }